refactor: slim backend_adapter/config.py to compatibility shim
