+++ uniqc/config.py
@@ -1,27 +1,351 @@
 """Project-wide configuration and platform credentials helpers.
 
-This module intentionally exposes the top-level API used by CLI and adapters:
-
-- Shared project config APIs (`load_config`, `save_config`, `get_platform_config`, ...)
-- Platform credential loaders used by cloud adapters (`load_originq_config`, ...)
+This module is the canonical source of truth for ``CONFIG_FILE`` and all
+configuration management functions (``load_config``, ``save_config``,
+``get_active_profile``, …).  ``uniqc.backend_adapter.config`` re-exports from here
+so that patching ``uniqc.config.CONFIG_FILE`` propagates to every import path.
 
 All credentials and cache settings are persisted in ``~/.uniqc/config.yaml``.
+
+Example configuration structure::
+
+    always_ai_hints: false
+    active_profile: default
+    default:
+      originq:
+        token: xxx
+      quafu:
+        token: xxx
+      quark:
+        QUARK_API_KEY: xxx
+      ibm:
+        token: xxx
+        proxy:
+          http: http://proxy:8080
+          https: https://proxy:8080
 """
 
+from __future__ import annotations
+
+import os
+from pathlib import Path
 from typing import Any
 
-from uniqc.backend_adapter.config import *  # noqa: F401,F403
+import yaml
 
+# ---------------------------------------------------------------------------
+# Top-level symbols (defined here so that uniqc.backend_adapter.config can
+# import them and both modules reference the same objects after patching)
+# ---------------------------------------------------------------------------
+
+CONFIG_DIR = Path.home() / ".uniqc"
+CONFIG_FILE = CONFIG_DIR / "config.yaml"
+
+DEFAULT_CONFIG: dict[str, Any] = {
+    "always_ai_hints": False,
+    "default": {
+        "originq": {
+            "token": "",
+            "available_qubits": [],
+            "available_topology": [],
+            "task_group_size": 200,
+        },
+        "quafu": {
+            "token": "",
+        },
+        "quark": {
+            "QUARK_API_KEY": "",
+        },
+        "ibm": {
+            "token": "",
+            "proxy": {
+                "http": "",
+                "https": "",
+            },
+        },
+    },
+}
+
+SUPPORTED_PLATFORMS = ["originq", "quafu", "quark", "ibm"]
+
+META_KEYS = frozenset({"active_profile", "always_ai_hints"})
+
+PLATFORM_REQUIRED_FIELDS = {
+    "originq": ["token"],
+    "quafu": ["token"],
+    "quark": ["QUARK_API_KEY"],
+    "ibm": ["token"],
+}
+
+
+# ---------------------------------------------------------------------------
+# Exceptions
+# ---------------------------------------------------------------------------
+
+class ConfigError(Exception):
+    pass
+
+
+class ConfigValidationError(ConfigError):
+    pass
+
+
+class PlatformNotFoundError(ConfigError):
+    pass
+
+
+class ProfileNotFoundError(ConfigError):
+    pass
+
+
+# ---------------------------------------------------------------------------
+# Core functions (also used by uniqc.backend_adapter.config via re-export)
+# ---------------------------------------------------------------------------
+
+def _ensure_config_dir() -> None:
+    CONFIG_DIR.mkdir(parents=True, exist_ok=True)
+
+
+def load_config(config_path: str | Path | None = None) -> dict[str, Any]:
+    path = Path(config_path) if config_path else CONFIG_FILE
+
+    if not path.exists():
+        return DEFAULT_CONFIG.copy()
+
+    try:
+        with open(path, encoding="utf-8") as f:
+            config = yaml.safe_load(f)
+        if config is None:
+            return DEFAULT_CONFIG.copy()
+        return config
+    except yaml.YAMLError as e:
+        raise ConfigError(f"Failed to parse YAML configuration: {e}") from e
+    except OSError as e:
+        raise ConfigError(f"Failed to read configuration file: {e}") from e
+
+
+def save_config(config: dict[str, Any], config_path: str | Path | None = None) -> None:
+    path = Path(config_path) if config_path else CONFIG_FILE
+    path.parent.mkdir(parents=True, exist_ok=True)
+
+    try:
+        with open(path, "w", encoding="utf-8") as f:
+            yaml.dump(
+                config,
+                f,
+                default_flow_style=False,
+                allow_unicode=True,
+                sort_keys=False,
+            )
+    except OSError as e:
+        raise ConfigError(f"Failed to write configuration file: {e}") from e
+
+
+def get_platform_config(
+    platform_name: str,
+    profile: str = "default",
+    config_path: str | Path | None = None,
+) -> dict[str, Any]:
+    if platform_name not in SUPPORTED_PLATFORMS:
+        raise PlatformNotFoundError(
+            f"Unsupported platform: {platform_name}. "
+            f"Supported platforms: {', '.join(SUPPORTED_PLATFORMS)}"
+        )
+
+    config = load_config(config_path)
+
+    if profile not in config:
+        raise ProfileNotFoundError(
+            f"Profile '{profile}' not found in configuration. "
+            f"Available profiles: {', '.join(config.keys())}"
+        )
+
+    profile_config = config[profile]
+
+    if platform_name not in profile_config:
+        raise ConfigError(
+            f"Platform '{platform_name}' not found in profile '{profile}'"
+        )
+
+    return profile_config[platform_name]
+
+
+def validate_config(
+    config: dict[str, Any] | None = None,
+    config_path: str | Path | None = None,
+) -> list[str]:
+    errors: list[str] = []
+
+    try:
+        cfg = config if config is not None else load_config(config_path)
+    except ConfigError as e:
+        return [str(e)]
+
+    if not isinstance(cfg, dict):
+        return ["Configuration must be a dictionary"]
+
+    if not cfg:
+        return ["Configuration is empty"]
+
+    for profile_name, profile_config in cfg.items():
+        if profile_name in META_KEYS:
+            continue
+        if not isinstance(profile_config, dict):
+            errors.append(f"Profile '{profile_name}' must be a dictionary")
+            continue
+
+        for platform_name in SUPPORTED_PLATFORMS:
+            if platform_name not in profile_config:
+                continue
+
+            platform_config = profile_config[platform_name]
+
+            if not isinstance(platform_config, dict):
+                errors.append(
+                    f"Platform '{platform_name}' in profile '{profile_name}' "
+                    "must be a dictionary"
+                )
+                continue
+
+            if platform_name == "quark":
+                if "QUARK_API_KEY" not in platform_config and "token" not in platform_config:
+                    errors.append(
+                        "Missing required field 'QUARK_API_KEY' for platform "
+                        f"'{platform_name}' in profile '{profile_name}'"
+                    )
+                continue
+
+            required_fields = PLATFORM_REQUIRED_FIELDS.get(platform_name, [])
+            for field in required_fields:
+                if field not in platform_config:
+                    errors.append(
+                        f"Missing required field '{field}' for platform "
+                        f"'{platform_name}' in profile '{profile_name}'"
+                    )
+
+            if platform_name == "ibm" and "proxy" in platform_config:
+                proxy = platform_config["proxy"]
+                if not isinstance(proxy, dict):
+                    errors.append(
+                        f"Proxy configuration for IBM in profile '{profile_name}' "
+                        "must be a dictionary"
+                    )
+                else:
+                    for proxy_type in ["http", "https"]:
+                        if proxy_type in proxy and not isinstance(proxy[proxy_type], str):
+                            errors.append(
+                                f"Proxy '{proxy_type}' for IBM in profile "
+                                f"'{profile_name}' must be a string"
+                            )
+
+    return errors
+
+
+def create_default_config(config_path: str | Path | None = None) -> None:
+    path = Path(config_path) if config_path else CONFIG_FILE
+    if path.exists():
+        return
+    _ensure_config_dir()
+    save_config(DEFAULT_CONFIG, path)
+
+
+def update_platform_config(
+    platform_name: str,
+    platform_config: dict[str, Any],
+    profile: str = "default",
+    config_path: str | Path | None = None,
+) -> None:
+    if platform_name not in SUPPORTED_PLATFORMS:
+        raise PlatformNotFoundError(
+            f"Unsupported platform: {platform_name}. "
+            f"Supported platforms: {', '.join(SUPPORTED_PLATFORMS)}"
+        )
+
+    config = load_config(config_path)
+
+    if profile not in config:
+        config[profile] = {}
+
+    config[profile][platform_name] = platform_config
+    save_config(config, config_path)
+
+
+def get_active_profile(config_path: str | Path | None = None) -> str:
+    env_profile = os.environ.get("UNIQC_PROFILE")
+    if env_profile:
+        return env_profile
+
+    config = load_config(config_path)
+    if "active_profile" in config:
+        return config["active_profile"]
+
+    return "default"
+
+
+def set_active_profile(
+    profile: str,
+    config_path: str | Path | None = None,
+) -> None:
+    config = load_config(config_path)
+
+    if profile not in config:
+        raise ProfileNotFoundError(
+            f"Profile '{profile}' not found in configuration. "
+            f"Available profiles: {', '.join(k for k in config if k not in META_KEYS)}"
+        )
+
+    config["active_profile"] = profile
+    save_config(config, config_path)
+
+
+def get_always_ai_hints(config_path: str | Path | None = None) -> bool:
+    config = load_config(config_path)
+    return bool(config.get("always_ai_hints", False))
+
+
+def set_always_ai_hints(
+    enabled: bool,
+    config_path: str | Path | None = None,
+) -> None:
+    config = load_config(config_path)
+    config["always_ai_hints"] = bool(enabled)
+    save_config(config, config_path)
+
+
+def get_originq_config(profile: str | None = None) -> dict[str, Any]:
+    if profile is None:
+        profile = get_active_profile()
+    return get_platform_config("originq", profile)
+
+
+def get_quafu_config(profile: str | None = None) -> dict[str, Any]:
+    if profile is None:
+        profile = get_active_profile()
+    return get_platform_config("quafu", profile)
+
+
+def get_quark_config(profile: str | None = None) -> dict[str, Any]:
+    if profile is None:
+        profile = get_active_profile()
+    return get_platform_config("quark", profile)
+
+
+def get_ibm_config(profile: str | None = None) -> dict[str, Any]:
+    if profile is None:
+        profile = get_active_profile()
+    return get_platform_config("ibm", profile)
+
+
+# ---------------------------------------------------------------------------
+# Platform-specific credential loaders
+# ---------------------------------------------------------------------------
 
 def _load_platform_config(platform: str) -> dict[str, Any]:
-    """Load ``platform`` config from the active YAML profile."""
-
     profile = get_active_profile()
     return get_platform_config(platform, profile)
 
 
 def load_originq_config() -> dict[str, Any]:
-    """Load OriginQ Cloud configuration from the active YAML profile."""
     config = _load_platform_config("originq")
     api_key = config.get("token", "") or None
 
@@ -39,7 +363,6 @@ def load_originq_config() -> dict[str, Any]:
 
 
 def load_quafu_config() -> dict[str, Any]:
-    """Load Quafu configuration from the active YAML profile."""
     config = _load_platform_config("quafu")
     api_token = config.get("token", "") or None
 
@@ -53,7 +376,6 @@ def load_quafu_config() -> dict[str, Any]:
 
 
 def load_quark_config() -> dict[str, Any]:
-    """Load QuarkStudio configuration from the active YAML profile."""
     config = _load_platform_config("quark")
     api_token = config.get("QUARK_API_KEY", "") or config.get("token", "") or None
 
@@ -67,7 +389,6 @@ def load_quark_config() -> dict[str, Any]:
 
 
 def load_ibm_config() -> dict[str, Any]:
-    """Load IBM Quantum configuration from the active YAML profile."""
     config = _load_platform_config("ibm")
     api_token = config.get("token", "") or None
 
@@ -81,7 +402,6 @@ def load_ibm_config() -> dict[str, Any]:
 
 
 def load_dummy_config() -> dict[str, Any]:
-    """Load OriginQ dummy execution config from the active YAML profile."""
     try:
         config = _load_platform_config("originq")
     except Exception:
